Сделал рефакторинг по SonarLint, коегде исправил синтаксические ошибк…

+++ src/main/java/ru/innopolis/stc9/service/GroupServiceImpl.java
@@ -2,7 +2,7 @@ package ru.innopolis.stc9.service;
 
 import ru.innopolis.stc9.dao.GroupDao;
 import ru.innopolis.stc9.dao.GroupDaoImpl;
-import ru.innopolis.stc9.pogo.Group;
+import ru.innopolis.stc9.pojo.Group;
 
 import java.util.List;
 
@@ -15,26 +15,31 @@ public class GroupServiceImpl implements GroupService {
 
     @Override
     public boolean addGroup(Group group) {
+        if (group == null) return false;
         return groupDao.addGroup(group);
     }
 
     @Override
     public boolean updateGroup(Group group) {
+        if (group == null) return false;
         return groupDao.updateGroup(group);
     }
 
     @Override
     public boolean deleteGroup(int groupId) {
+        if (groupId < 0) return false;
         return groupDao.deleteGroup(groupId);
     }
 
     @Override
     public Group findGroupById(int id) {
+        if (id < 0) return null;
         return groupDao.findGroupById(id);
     }
 
     @Override
     public Group findGroupByName(String name) {
+        if ((name == null) || name.isEmpty()) return null;
         return groupDao.findGroupByName(name);
     }
